+++ CHANGES.rst
@@ -28,6 +28,9 @@ resample
   results in modified values in the resampled images. New computations
   significantly reduce photometric errors. [#7894]
 
+- Apply a median filter to IVM weight array to better handle saturated
+  pixels. [#7563]
+
 
 1.12.5 (2023-10-19)
 ===================
@@ -353,6 +356,7 @@ resample
 
 - Update the following exposure time keywords: XPOSURE (EFFEXPTM),
   DURATION and TELAPSE. [#7793]
+
 
 residual_fringe
 ---------------